backend/requirements.txt
# backend/requirements.txt

# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# AI/ML libraries
transformers==4.35.2
torch==2.1.1
torchaudio==2.1.1
openai-whisper==20231117
numpy==1.25.2
scipy==1.11.4

# Audio processing
sounddevice==0.4.6

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0